refactor(popup): extract utility functions to popup/utils.ts
Move escapeHtml, getScoreColor, and exportResults out of the main
popup.ts entry point into a dedicated utils module. popup.ts re-exports
them to keep the public API unchanged.
